break CryptoAlg into component algebras
